[Feature] Refactor code to pass tests

+++ server/api/controllers/documentController.js
@@ -11,14 +11,23 @@ module.exports = {
         const offset = req.query.offset || 5;
         const allDocuments = await Document.find({});
         let documents = await allDocuments.filter((document, user) => documentAccessCheck(document, req.user));
-        res.status(200).json(documents);
+        res.status(200).json({ success: true, documents: documents });
     },
     // Get a single document
     getDocument: async (req, res, next) => {
         const { documentID } = req.value.params;
         const document = await Document.findById(documentID);
-        const doc = await documentAccessCheck(document, req.user) || {message: "Document does not exist or you are not allowed to access it"};
-        res.status(200).json(doc);
+        if (document){
+            const doc = await documentAccessCheck(document, req.user);
+            if (doc) {
+                res.status(200).json({ success: true, document: doc });
+            } else {
+                res.status(401).json({ success: false, message: "Document does not exist or you are not allowed to access it" });
+            }
+        } else {
+            res.status(404).json({ success: false, message: 'Document does not exist or you do not have access to it' });
+        }
+        
     },
     // Create a document
     createDocument: async (req, res, next) => {
@@ -33,7 +42,7 @@ module.exports = {
         await owner.documents.push(newDocument);
         // Save the owner
         await owner.save();
-        res.status(201).json(newDocument);
+        res.status(201).json({ success: true, document: newDocument, message: 'Document successfully created'});
     },
     // Update a document (PATCH)
     updateDocument: async (req, res, next) => {
@@ -41,12 +50,17 @@ module.exports = {
         const newDocument = req.value.body;
         // Find the document and check access
         const document = await Document.findById(documentID);
-        const doc = await documentRoleCheck(document, req.user);
-        if (!doc) {
-            res.status(400).json({message: "Document does not exist or you are not allowed to modify it"});
+        if (document){
+            const doc = await documentRoleCheck(document, req.user);
+            if (doc) {
+                const result = await Document.findByIdAndUpdate(documentID, newDocument, {new: true});
+                res.status(200).json({success: true, message: 'Document updated successfully'}); 
+            } else {
+                res.status(403).json({success: false, message: "Document does not exist or you are not allowed to modify it"});
+            }
+        } else {
+            res.status(404).json({ success: false, message: 'Document does not exist or you do not have access to it' });
         }
-        const result = await Document.findByIdAndUpdate(documentID, newDocument, {new: true});
-        res.status(200).json({success: true});        
     },
     // Replace a document (PUT)
     replaceDocument: async (req, res, next) => {
@@ -54,37 +68,50 @@ module.exports = {
         const newDocument = req.value.body;
         // Find the document and check access
         const document = await Document.findById(documentID);
-        const doc = await documentRoleCheck(document, req.user);
-        if (!doc) {
-            res.status(400).json({message: "Document does not exist or you are not allowed to modify it"});
-        }
-        const result = await Document.findByIdAndUpdate(documentID, newDocument, {new: true});
-        res.status(200).json({success: true});             
+        if (document){
+            const doc = await documentRoleCheck(document, req.user);
+            if (!doc) {
+                res.status(403).json({success: false, message: "Document does not exist or you are not allowed to modify it"});
+            } else if (doc) {
+                const result = await Document.findByIdAndUpdate(documentID, newDocument, {new: true});
+                res.status(200).json({success: true, message: 'Document updated successfully'}); 
+            }
+        } else {
+            res.status(404).json({ success: false, message: 'Document does not exist or you do not have access to it' });
+        }         
     },
     // Delete a document
     deleteDocument: async (req, res, next) => {
         const { documentID } = req.value.params;
         // Find the document and check access
         const document = await Document.findById(documentID);
-        const doc = await documentRoleCheck(document, req.user);
-        if (!doc) {
-            res.status(400).json({message: "Document does not exist or you are not allowed to modify it"});
-        }        
-        const result = await Document.findByIdAndRemove(documentID);
-        // Remove from owner's list of documents
-        const owner = await User.findById(req.user.id);
-        // owner.documents.pull(document)
-        owner.documents = await owner.documents.filter((document) => { 
-            return document != documentID;
-        });
-        await owner.save()
-        // remember to remove from a user's list
-        res.status(200).json({success: true});
+        if (document){
+            const doc = await documentRoleCheck(document, req.user);
+            if (!doc) {
+                res.status(403).json({success: false, message: "Document does not exist or you are not allowed to modify it"});
+            }        
+            const result = await Document.findByIdAndRemove(documentID);
+            // Remove from owner's list of documents
+            const owner = await User.findById(req.user.id);
+            // owner.documents.pull(document)
+            owner.documents = await owner.documents.filter((document) => { 
+                return document != documentID;
+            });
+            await owner.save()
+            // remember to remove from a user's list
+            res.status(200).json({ success: true, message: 'Document deleted successfully'});
+        } else {
+            res.status(404).json({ success: false, message: 'Document does not exist or you do not have access to it' });
+        }
     },
     // Search for a document
     searchDocument: async (req, res, next) => {
         const query = req.query.q;
-        const allDocuments = Document.find({ title: query});
+        const allDocuments = Document.find({    
+            "$text": {
+                "$search":query
+            }
+        });
         let documents = await allDocuments.filter((document, user) => documentAccessCheck(document, req.user));
         res.status(200).json(documents);
     }
